Changes according CR, add mocks
