+++ libs/shared/data/features/feedback/index.ts
@@ -1,1 +1,2 @@
+export * from "./mocks/items";
 export * from "./sample";
