+++ libs/shared/src/constants/results-limits.ts
@@ -1,5 +1,5 @@
 export const RESULTS_LIMITS = {
   limit: 5,
   offset: 0,
-  minMatch: 3
+  minMatch: 2
 };
