+++ libs/shared/src/utils/search-handler.ts
@@ -1,20 +1,6 @@
 import Fuse from "fuse.js";
 
-import { RESULTS_LIMITS } from "../constants";
-
-export interface ItemsType {
-  title: string;
-  id: string;
-}
-
-export const items: ItemsType[] = [
-  { title: "Advanced typescript", id: "8ef796d1-b39d-4c4b-91a4-689534944d66" },
-  { title: "Javascript for begginers", id: "3d7812ee-43b5-43f3-8a3e-ea8938cbf8c1" },
-  { title: "Advanced React", id: "db5d7672-b143-4fa4-941c-5aa71e0eb575" },
-  { title: "Advanced Javascript", id: "1523a989-ceee-422a-b460-d5e2b842f742" },
-  { title: "Typescript for begginers", id: "acf31308-3432-4b08-a807-538d4f8e288d" },
-  { title: "React for begginers", id: "73bae9d6-3af9-4927-998b-605fc215dfac" }
-];
+import { items, ItemsType } from "../../data/features/feedback/mocks/items";
 
 interface SearchHandlerConfig {
   key: string;
@@ -41,7 +27,6 @@ export const searchHandler: SearchHandlerType = ({ key, text, offset, limit }) =
     items.map((item, index) => ({
       item,
       matches: [],
-      minMatchCharLength: RESULTS_LIMITS.minMatch,
       refIndex: index
     })),
     offset,
